ai-model/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2

# Agentic AI Dependencies
# Updated langchain packages to resolve dependency conflicts
langgraph==0.0.40
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-ollama==0.0.1
ollama==0.5.0

# Vector Database & RAG
chromadb==0.4.15
faiss-cpu==1.7.4
sentence-transformers==2.2.0

# Enhanced NLP & AI
transformers==4.30.0
torch==2.0.0

# Additional utilities
numpy==1.24.0
pandas==2.0.3